restructuration 1.0
