+++ percephone/analysis/decoding.py
@@ -1,0 +1,5 @@
+"""
+Théo Gauvrit 18/01/2024
+decoding of the behaviors labels from neurons activity
+"""
+
