+++ percephone/plts/style.py
@@ -1,0 +1,4 @@
+"""
+Théo Gauvrit 18/01/2024
+Style and aesthetics for plots
+"""
