+++ percephone/plts/utils.py
@@ -1,0 +1,4 @@
+"""
+Théo Gauvrit 18/01/2024
+Utility function for plots
+"""
